+++ kronos/__init__.py
@@ -1,13 +1,20 @@
 from functools import wraps
 
+import django
 from django.core.management import get_commands, load_command_class
-from django.utils.importlib import import_module
-from kronos.settings import PROJECT_MODULE_NAME, KRONOS_PYTHON, KRONOS_MANAGE, \
-    KRONOS_PYTHONPATH, KRONOS_POSTFIX
+
+try:
+    from importlib import import_module
+except ImportError:
+    from django.utils.importlib import import_module
+
+from kronos.settings import PROJECT_MODULE, KRONOS_PYTHON, KRONOS_MANAGE, \
+    KRONOS_PYTHONPATH, KRONOS_POSTFIX, KRONOS_PREFIX
 from django.conf import settings
 from kronos.utils import read_crontab, write_crontab, delete_crontab
 from kronos.version import __version__
-
+import six
+from django.utils.module_loading import autodiscover_modules
 
 tasks = []
 
@@ -16,42 +23,58 @@ def load():
     """
     Load ``cron`` modules for applications listed in ``INSTALLED_APPS``.
     """
-    if PROJECT_MODULE_NAME:
-        paths = ['%s.cron' % PROJECT_MODULE_NAME]
-        if '.' in PROJECT_MODULE_NAME:
-            paths.append('%s.cron' % '.'.join(
-                PROJECT_MODULE_NAME.split('.')[0:-1]))
-    else:
-        paths = []
+    autodiscover_modules('cron')
 
-    for application in settings.INSTALLED_APPS:
-        paths.append('%s.cron' % application)
-
-    # load kronostasks
-    for p in paths:
+    if '.' in PROJECT_MODULE.__name__:
         try:
-            import_module(p)
-        except ImportError:
-            pass
+            import_module('%s.cron' % '.'.join(
+	            PROJECT_MODULE.__name__.split('.')[0:-1]))
+        except ImportError as e:
+            if 'No module named' not in str(e):
+                print(e)
 
     # load django tasks
     for cmd, app in get_commands().items():
-        load_command_class(app, cmd)
+        try:
+            load_command_class(app, cmd)
+        except django.core.exceptions.ImproperlyConfigured:
+            pass
 
 
-def register(schedule):
+def register(schedule, *args, **kwargs):
     def decorator(function):
         global tasks
+        passed_args = []
+
+        if "args" in kwargs:
+            for key, value in six.iteritems(kwargs["args"]):
+                if isinstance(value, dict):
+                    raise TypeError('Parse for dict arguments not yet implemented.')
+
+                if isinstance(value, list):
+                    temp_args = ",".join(map(str, value))
+                    passed_args.append("{}={}".format(key, temp_args))
+                else:
+                    if value is None:
+                        arg_text = "{}"
+                    elif isinstance(value, str):
+                        arg_text = '{} "{}"'
+                    else:
+                        arg_text = '{} {}'
+
+                    passed_args.append(arg_text.format(key, value))
 
         if hasattr(function, 'handle'):
             # django command
-            function.cron_expression = '%(schedule)s %(python)s %(manage)s ' \
-                '%(task)s --settings=%(settings_module)s %(postfix)s' \
+            function.cron_expression = '%(schedule)s %(prefix)s %(python)s %(manage)s ' \
+                '%(task)s %(passed_args)s --settings=%(settings_module)s %(postfix)s' \
                 '$KRONOS_BREAD_CRUMB' % {
                     'schedule': schedule,
+                    'prefix': KRONOS_PREFIX,
                     'python': KRONOS_PYTHON,
                     'manage': KRONOS_MANAGE,
                     'task': function.__module__.split('.')[-1],
+                    'passed_args': " ".join(passed_args),
                     'settings_module': settings.SETTINGS_MODULE,
                     'postfix': KRONOS_POSTFIX
                 }
@@ -59,13 +82,15 @@ def register(schedule):
                 django_command=True,
                 fn=function)
         else:
-            function.cron_expression = '%(schedule)s %(python)s %(manage)s ' \
-                'runtask %(task)s --settings=%(settings_module)s ' \
+            function.cron_expression = '%(schedule)s %(prefix)s %(python)s %(manage)s ' \
+                'runtask %(task)s %(passed_args)s --settings=%(settings_module)s ' \
                 '%(postfix)s $KRONOS_BREAD_CRUMB' % {
                     'schedule': schedule,
+                    'prefix': KRONOS_PREFIX,
                     'python': KRONOS_PYTHON,
                     'manage': KRONOS_MANAGE,
                     'task': function.__name__,
+                    'passed_args': " ".join(passed_args),
                     'settings_module': settings.SETTINGS_MODULE,
                     'postfix': KRONOS_POSTFIX
                 }
@@ -91,7 +116,7 @@ def install():
     Register tasks with cron.
     """
     load()
-    current_crontab = read_crontab()
+    current_crontab = six.u(read_crontab())
 
     new_crontab = ''
     for task in tasks:
@@ -118,7 +143,7 @@ def uninstall():
     current_crontab = read_crontab()
 
     new_crontab = ''
-    for line in current_crontab.split('\n')[:-1]:
+    for line in six.u(current_crontab).split('\n')[:-1]:
         exp = '%(python)s %(manage)s runtask' % {
             'python': KRONOS_PYTHON,
             'manage': KRONOS_MANAGE,